Widen small integers only to Int32, not Int
Reasons, as described in #14407:
- 32-bit integer division is significantly faster than 64-bit integer division.
- Consistency across platforms (widen(Int(8)) will do the same thing on 32 & 64 bit)
- Smaller integers can have larger gains from vectorization

Changes:
- Widen small integers only to Int32, not Int
- Ensure that `widen` tests actually check the result type, not just the value.
- Update tests.
- No documentation update necessary since the actual return type is left unspecified.

Closes #14407.

+++ test/reduce.jl
@@ -33,7 +33,7 @@
 
 # sum
 
-@test sum(Int8[]) === 0
+@test sum(Int8[]) === Int32(0)
 @test sum(Int[]) === Int(0)
 @test sum(Float64[]) === 0.0
 
@@ -41,7 +41,7 @@
 @test sum(3) === 3
 @test sum(3.0) === 3.0
 
-@test sum([Int8(3)]) === 3
+@test sum([Int8(3)]) === Int32(3)
 @test sum([3]) === 3
 @test sum([3.0]) === 3.0
 
@@ -62,14 +62,14 @@ fz = float(z)
 a = randn(32) # need >16 elements to trigger BLAS code path
 b = complex(randn(32), randn(32))
 @test sumabs(Float64[]) === 0.0
-@test sumabs([Int8(-2)]) === 2
+@test sumabs([Int8(-2)]) === Int32(2)
 @test sumabs(z) === 14
 @test sumabs(fz) === 14.0
 @test_approx_eq sumabs(a) sum(abs(a))
 @test_approx_eq sumabs(b) sum(abs(b))
 
 @test sumabs2(Float64[]) === 0.0
-@test sumabs2([Int8(-2)]) === 4
+@test sumabs2([Int8(-2)]) === Int32(4)
 @test sumabs2(z) === 54
 @test sumabs2(fz) === 54.0
 @test_approx_eq sumabs2(a) sum(abs2(a))
@@ -105,11 +105,11 @@ end
 # prod
 
 @test prod(Int[]) === 1
-@test prod(Int8[]) === 1
+@test prod(Int8[]) === Int32(1)
 @test prod(Float64[]) === 1.0
 
 @test prod([3]) === 3
-@test prod([Int8(3)]) === 3
+@test prod([Int8(3)]) === Int32(3)
 @test prod([3.0]) === 3.0
 
 @test prod(z) === 120
@@ -124,8 +124,8 @@ end
 prod2(itr) = invoke(prod, Tuple{Any}, itr)
 @test prod(Int[]) === prod2(Int[]) === 1
 @test prod(Int[7]) === prod2(Int[7]) === 7
-@test typeof(prod(Int8[])) == typeof(prod(Int8[1])) == typeof(prod(Int8[1, 7])) == Int
-@test typeof(prod2(Int8[])) == typeof(prod2(Int8[1])) == typeof(prod2(Int8[1 7])) == Int
+@test typeof(prod(Int8[])) == typeof(prod(Int8[1])) == typeof(prod(Int8[1, 7])) == Int32
+@test typeof(prod2(Int8[])) == typeof(prod2(Int8[1])) == typeof(prod2(Int8[1 7])) == Int32
 
 # maximum & minimum & extrema
